# p751: speed up montgomery reduction with mulx/adox