In [this paper](https://eprint.iacr.org/2017/1015.pdf) a technique was described to improve the performance of Montgomery reduction for Montgomery-friendly moduli. This adds an implementation using the mulx, adox and adcx instructions, available in the BMI2 (since Haswell) and ADX (since Broadwell) instruction set extensions. The decision which to pick is now made at runtime.

Before, on my Kaby Lake laptop:
```
BenchmarkThreePointLadder379BitScalar-8   	    1000	   1965435 ns/op
BenchmarkR2L379BitScalar-8                	    1000	   1939736 ns/op
BenchmarkExtensionFieldElementMul-8       	 3000000	       425 ns/op
BenchmarkExtensionFieldElementInv-8       	   10000	    125856 ns/op
BenchmarkExtensionFieldElementSquare-8    	 5000000	       317 ns/op
BenchmarkExtensionFieldElementAdd-8       	50000000	        28.8 ns/op
BenchmarkExtensionFieldElementSub-8       	50000000	        25.6 ns/op
BenchmarkPrimeFieldElementMul-8           	10000000	       144 ns/op
BenchmarkPrimeFieldElementInv-8           	   10000	    128079 ns/op
BenchmarkPrimeFieldElementSqrt-8          	   10000	    124538 ns/op
BenchmarkPrimeFieldElementSquare-8        	10000000	       135 ns/op
BenchmarkPrimeFieldElementAdd-8           	100000000	        15.4 ns/op
BenchmarkPrimeFieldElementSub-8           	100000000	        14.9 ns/op
BenchmarkFp751Multiply-8                  	20000000	        78.5 ns/op
BenchmarkFp751MontgomeryReduce-8          	30000000	        59.3 ns/op
BenchmarkFp751AddReduced-8                	100000000	        12.7 ns/op
BenchmarkFp751SubReduced-8                	100000000	        11.8 ns/op
BenchmarkFp751ConditionalSwap-8           	100000000	        22.7 ns/op
BenchmarkFp751StrongReduce-8              	100000000	        12.6 ns/op
BenchmarkFp751AddLazy-8                   	200000000	         8.08 ns/op
BenchmarkFp751X2AddLazy-8                 	100000000	        11.9 ns/op
BenchmarkFp751X2SubLazy-8                 	100000000	        17.6 ns/op
```

The faster algorithm with BMI2 (mulx):
```
BenchmarkThreePointLadder379BitScalar-8   	    1000	   1842450 ns/op
BenchmarkR2L379BitScalar-8                	    1000	   1863163 ns/op
BenchmarkExtensionFieldElementMul-8       	 3000000	       408 ns/op
BenchmarkExtensionFieldElementInv-8       	   10000	    119046 ns/op
BenchmarkExtensionFieldElementSquare-8    	 5000000	       297 ns/op
BenchmarkExtensionFieldElementAdd-8       	50000000	        28.3 ns/op
BenchmarkExtensionFieldElementSub-8       	50000000	        25.7 ns/op
BenchmarkPrimeFieldElementMul-8           	10000000	       132 ns/op
BenchmarkPrimeFieldElementInv-8           	   10000	    117291 ns/op
BenchmarkPrimeFieldElementSqrt-8          	   10000	    116817 ns/op
BenchmarkPrimeFieldElementSquare-8        	10000000	       132 ns/op
BenchmarkPrimeFieldElementAdd-8           	100000000	        15.5 ns/op
BenchmarkPrimeFieldElementSub-8           	100000000	        13.9 ns/op
BenchmarkFp751Multiply-8                  	20000000	        74.1 ns/op
BenchmarkFp751MontgomeryReduce-8          	30000000	        46.6 ns/op
BenchmarkFp751AddReduced-8                	100000000	        12.9 ns/op
BenchmarkFp751SubReduced-8                	100000000	        12.0 ns/op
BenchmarkFp751ConditionalSwap-8           	100000000	        23.3 ns/op
BenchmarkFp751StrongReduce-8              	100000000	        13.1 ns/op
BenchmarkFp751AddLazy-8                   	200000000	         7.64 ns/op
BenchmarkFp751X2AddLazy-8                 	100000000	        11.0 ns/op
BenchmarkFp751X2SubLazy-8                 	100000000	        17.3 ns/op
```

The faster algorithm with BMI2 (mulx) and ADX (adcx, adox):
```
BenchmarkThreePointLadder379BitScalar-8   	    1000	   1834232 ns/op
BenchmarkR2L379BitScalar-8                	    1000	   1891392 ns/op
BenchmarkExtensionFieldElementMul-8       	 3000000	       429 ns/op
BenchmarkExtensionFieldElementInv-8       	   10000	    115947 ns/op
BenchmarkExtensionFieldElementSquare-8    	 5000000	       287 ns/op
BenchmarkExtensionFieldElementAdd-8       	50000000	        28.6 ns/op
BenchmarkExtensionFieldElementSub-8       	50000000	        26.6 ns/op
BenchmarkPrimeFieldElementMul-8           	10000000	       124 ns/op
BenchmarkPrimeFieldElementInv-8           	   10000	    115491 ns/op
BenchmarkPrimeFieldElementSqrt-8          	   10000	    111315 ns/op
BenchmarkPrimeFieldElementSquare-8        	10000000	       123 ns/op
BenchmarkPrimeFieldElementAdd-8           	100000000	        15.4 ns/op
BenchmarkPrimeFieldElementSub-8           	100000000	        14.1 ns/op
BenchmarkFp751Multiply-8                  	20000000	        75.5 ns/op
BenchmarkFp751MontgomeryReduce-8          	30000000	        40.4 ns/op
BenchmarkFp751AddReduced-8                	100000000	        12.8 ns/op
BenchmarkFp751SubReduced-8                	100000000	        11.7 ns/op
BenchmarkFp751ConditionalSwap-8           	50000000	        22.7 ns/op
BenchmarkFp751StrongReduce-8              	100000000	        12.8 ns/op
BenchmarkFp751AddLazy-8                   	200000000	         7.53 ns/op
BenchmarkFp751X2AddLazy-8                 	100000000	        11.3 ns/op
BenchmarkFp751X2SubLazy-8                 	100000000	        17.0 ns/op
```