@henrydcase 2018-09-03
processor   : 0
vendor_id   : GenuineIntel
cpu family  : 6
model       : 142
model name  : Intel(R) Core(TM) i7-7600U CPU @ 2.80GHz
stepping    : 9
microcode   : 0x84
cpu MHz     : 3858.669

> export GOMAXPROCS=1

New:
```
> ./sidh.test -test.v -test.bench="BenchmarkBobKeyGenPub" -test.run="^A" -test.benchtime=20s -test.cpu 1
goos: linux
goarch: amd64
pkg: github.com/cloudflare/p751sidh/sidh
BenchmarkBobKeyGenPub       2000      13087326 ns/op
> ./sidh.test -test.v -test.bench="BenchmarkSharedSecretBob" -test.run="^A" -test.benchtime=20s -test.cpu 1
goos: linux
goarch: amd64
pkg: github.com/cloudflare/p751sidh/sidh
BenchmarkSharedSecretBob        3000      11307942 ns/op
```


Old:
```
> ./sidh.test -test.v -test.bench="BenchmarkBobKeyGenPub" -test.run="^A" -test.benchtime=20s -test.cpu 1
goos: linux
goarch: amd64
pkg: github.com/cloudflare/p751sidh/sidh
BenchmarkBobKeyGenPub       2000      13148866 ns/op
> ./sidh.test -test.v -test.bench="BenchmarkSharedSecretBob" -test.run="^A" -test.benchtime=20s -test.cpu 1
goos: linux
goarch: amd64
pkg: github.com/cloudflare/p751sidh/sidh
BenchmarkSharedSecretBob        3000       9986017 ns/op
PASS

```