@Ko- 2018-09-03
Should that be reversed or did it actually get slower on that machine?