@henrydcase 2018-09-04
Old:
```
BenchmarkFp751MontgomeryReduce 	20000000	        60.0 ns/op
```
New:
```
BenchmarkFp751MontgomeryReduce 	30000000	        43.9 ns/op
```