@henrydcase 2018-09-04
I think patch improves redc, but doesn't impact overall perf. Still that's OK