@henrydcase 2018-09-13
Changes retested in  ``review/amd64-optim-montgomeryReduce``. 
Work OK,